backend/requirements.txt

fastapi==0.95.2
uvicorn[standard]==0.21.1
sqlalchemy==1.4.49
asyncpg==0.27.0
databases==0.7.0
pydantic==1.10.12
python-jose==3.3.0
passlib==1.7.4
alembic==1.11.1

kafka-python==2.0.2
scikit-learn==1.3.2
pandas==2.1.2
joblib==1.3.2
